GraphQL 4 compat
