+++ src/models/Link.php
@@ -15,11 +15,8 @@ use SilverStripe\Forms\TreeDropdownField;
 use SilverStripe\ORM\DataObject;
 use SilverStripe\ORM\ValidationResult;
 use SilverStripe\Control\Director;
-use SilverStripe\View\SSViewer;
 use SilverStripe\CMS\Controllers\ContentController;
 use UncleCheese\DisplayLogic\Forms\Wrapper;
-use SilverStripe\GraphQL\Scaffolding\Interfaces\ScaffoldingProvider;
-use SilverStripe\GraphQL\Scaffolding\Scaffolders\SchemaScaffolder;
 use SilverStripe\Assets\Folder;
 
 /**
@@ -28,8 +25,7 @@ use SilverStripe\Assets\Folder;
  * @package silverstripe
  * @subpackage silverstripe-link
  */
-class Link extends DataObject implements
-    ScaffoldingProvider
+class Link extends DataObject
 {
     /**
      * Defines the database table name
@@ -393,37 +389,6 @@ class Link extends DataObject implements
                     break;
             }
         }
-    }
-
-    public function provideGraphQLScaffolding(SchemaScaffolder $scaffolder)
-    {
-        $type = $scaffolder->type($this->ClassName);
-
-        $type->addAllFields()
-            ->addFields($this->gqlFields())
-            ->operation(SchemaScaffolder::READ)
-                ->setName('readLinks')
-                ->setUsePagination(false)
-                ->end()
-            ->operation(SchemaScaffolder::READ_ONE)
-                ->setName('readOneLink')
-                ->end()
-            ->operation(SchemaScaffolder::CREATE)
-                ->setName('createLink')
-                ->end()
-            ->operation(SchemaScaffolder::UPDATE)
-                ->setName('updateLink')
-                ->end()
-            ->operation(SchemaScaffolder::DELETE)
-                ->setName('deleteLink')
-                ->end()
-            ->end();
-        foreach ($this->gqlNestedQueries() as $query => $paginated) {
-            $type->nestedQuery($query)
-                ->setUsePagination($paginated)
-                ->end();
-        }
-        return $scaffolder;
     }
 
     /**
